Deprecated AbstractBlockServiceTestCase class with BlockServiceTestCase
