+++ tests/Block/Service/MenuBlockServiceTest.php
@@ -16,14 +16,14 @@ namespace Sonata\BlockBundle\Tests\Block\Service;
 use Knp\Menu\Provider\MenuProviderInterface;
 use Sonata\BlockBundle\Block\Service\MenuBlockService;
 use Sonata\BlockBundle\Menu\MenuRegistryInterface;
-use Sonata\BlockBundle\Test\AbstractBlockServiceTestCase;
+use Sonata\BlockBundle\Test\BlockServiceTestCase;
 use Sonata\Form\Type\ImmutableArrayType;
 use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\TextType;
 use Symfony\Component\Form\FormTypeInterface;
 
-final class MenuBlockServiceTest extends AbstractBlockServiceTestCase
+final class MenuBlockServiceTest extends BlockServiceTestCase
 {
     /**
      * @var MenuProviderInterface
