+++ tests/Block/Service/TextBlockServiceTest.php
@@ -16,10 +16,10 @@ namespace Sonata\BlockBundle\Tests\Block\Service;
 use Sonata\BlockBundle\Block\BlockContext;
 use Sonata\BlockBundle\Block\Service\TextBlockService;
 use Sonata\BlockBundle\Model\Block;
-use Sonata\BlockBundle\Test\AbstractBlockServiceTestCase;
+use Sonata\BlockBundle\Test\BlockServiceTestCase;
 use Sonata\BlockBundle\Util\OptionsResolver;
 
-final class TextBlockServiceTest extends AbstractBlockServiceTestCase
+final class TextBlockServiceTest extends BlockServiceTestCase
 {
     public function testService()
     {
